a little fix
declaration of return value int to NSInteger

+++ spotlist/ViewController.m
@@ -101,7 +101,7 @@
     
     // エラーコードをログに出力
     NSInteger errorCode = [[[jsonDic objectForKey:@"meta"] objectForKey:@"code"] integerValue];
-    NSLog(@"errorCode = %d", errorCode);
+    NSLog(@"errorCode = %ld", (long)errorCode);
     
     // 結果取得
     NSArray *venues = [[jsonDic objectForKey:@"response"] objectForKey:@"venues"];
@@ -111,15 +111,15 @@
 }
 
 // The number of columns of data
-- (int)numberOfComponentsInPickerView:(UIPickerView *)pickerView
+- (NSInteger)numberOfComponentsInPickerView:(UIPickerView *)pickerView
 {
     return 1;
 }
 
 // The number of rows of data
-- (int)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
+- (NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
 {
-    return _pickerData.count;
+    return (int)_pickerData.count;
 }
 
 // The data to return for the row and component (column) that's being passed in
